Migrate AnimationChain.MonoGame/KNI onto AnimationEditorCommon (#934 phase 2)
Deletes the package's duplicated .achx POCO layer and non-generic runtime types
(AnimationChain, AnimationChainList, AnimationPlayer, AnimationShapeFrame,
ColorOperation) in favor of AnimationEditorCommon's generic AnimationChain<TFrame>/
AnimationChainList<TFrame>/AnimationPlayer<TFrame> closed over a new
AnimationFrame : AnimationFrameBase<Texture2D>. AchxLoader and a new
AnimationChainListSaveExtensions (ToAnimationChainList/TryReload) carry the
package-specific bits Common can't own: sticky per-frame color resolution and
PixelRectangle->Rectangle conversion (PixelRectangleExtensions.ToXnaRectangle).

TryReload is named to avoid AnimationChainList<TFrame>'s own instance
TryReloadFrom overloads -- same name + compatible arg count let the instance
method silently shadow an extension with a frame-factory that always returns
null, dropping every frame on reload (caught via the existing test suite).

Both AnimationChain.MonoGame.csproj and AnimationChain.KNI.csproj now
ProjectReference AnimationEditorCommon; publish-animationchain-packages.yml
publishes it first, in lockstep, mirroring publish.yml's existing treatment
for FlatRedBall2.MonoGame/Kni.
