+++ src/AnimationChain.MonoGame/Runtime/AnimationFrameColor.cs
@@ -1,3 +1,5 @@
+using FlatRedBall2.Animation;
+using FlatRedBall2.AnimationEditorCommon;
 using Microsoft.Xna.Framework;
 using Microsoft.Xna.Framework.Graphics;
 
@@ -13,15 +15,15 @@ public static class AnimationFrameColor
     /// Combines <paramref name="frame"/>'s color channels with <paramref name="baseColor"/> and
     /// returns the resulting <see cref="Color"/> to pass to <see cref="SpriteBatch"/>'s draw call.
     /// <para>
-    /// <see cref="ColorOperation.Multiply"/> scales R/G/B by <see cref="AnimationFrame.Red"/>/
-    /// <see cref="AnimationFrame.Green"/>/<see cref="AnimationFrame.Blue"/> (unset channels default
-    /// to 255, the identity). <see cref="ColorOperation.Add"/> leaves RGB unchanged here — it is
-    /// applied by a pixel shader instead, since <see cref="SpriteBatch"/> can only multiply, not
-    /// offset, texture color. See <see cref="GetAddOffset"/> and
+    /// <see cref="ColorOperation.Multiply"/> scales R/G/B by <see cref="AnimationFrameBase.Red"/>/
+    /// <see cref="AnimationFrameBase.Green"/>/<see cref="AnimationFrameBase.Blue"/> (unset channels
+    /// default to 255, the identity). <see cref="ColorOperation.Add"/> leaves RGB unchanged here —
+    /// it is applied by a pixel shader instead, since <see cref="SpriteBatch"/> can only multiply,
+    /// not offset, texture color. See <see cref="GetAddOffset"/> and
     /// <see cref="SpriteBatchExtensions.DrawAnimation"/>.
     /// </para>
     /// <para>
-    /// <see cref="AnimationFrame.Alpha"/> is independent of <see cref="AnimationFrame.ColorOperation"/>
+    /// <see cref="AnimationFrameBase.Alpha"/> is independent of <see cref="AnimationFrameBase.ColorOperation"/>
     /// and always scales the base color's alpha when set.
     /// </para>
     /// </summary>
@@ -50,7 +52,7 @@ public static class AnimationFrameColor
 
     /// <summary>
     /// Computes the per-channel offset (range -1..1) that <see cref="SpriteBatchExtensions.DrawAnimation"/>
-    /// passes to the add-color pixel shader for a frame whose <see cref="AnimationFrame.ColorOperation"/>
+    /// passes to the add-color pixel shader for a frame whose <see cref="AnimationFrameBase.ColorOperation"/>
     /// is <see cref="ColorOperation.Add"/>. Unset channels default to 0 (the identity for Add). Values
     /// are clamped to the Animation Editor's authored range (-255..255) before scaling.
     /// </summary>
